Add dependabot to manage dependency updates.
This will cover python, npm and github actions.

The python and npm dependencies will be grouped together. We have a number
of dependencies already, so I'd like to avoid flooding our PRs with dependency
updates. The GitHub actions are fairly small, so I suspect we can have those
broken out.