Add basic playwright test
This adds two tests:
- Smoke test of web app
- Smoke test of wagtail application

This also adds a local.sh helper script to run runserver and tailwind start in one file.
It's not usable by windows folks, but it should provide a guide to implementing
one for windows.

The testing auth flow is to have a predefined username and password for the
playwright testing user. This means if you want to run the smoke tests against
production locally, you'll need the password for the production playwright
user. This is a bit risky considering the test user has Editor permissions to
access the wagtail admin.