Upgrades our npm dependencies and tailwind specifically to v4
This uncovered a few issues that I don't know when they changed in tailwind,
but they needed to be fixed. I suspect the column size issue actually came
from the bootstrap to tailwind port, but I'm not sure.

I don't like that a lot of stuff got pushed into djangonaut-space.css out of
the components layer, but I couldn't figure it out.