Limit the access of Session Organizers in the admin.
This creates a auth.Group for session organizers which determines which
areas of the admin a user can access. This is largely view and add permissions.
There are some edit rights. Very few cases can a user delete data.

This also adds people to the Session Organizers group when they are made
a session organizer via SessionMembership instance. This grants them access
to the admin as well. If the controlling user is a super user, it reminds them
to check on the members of that group to remove anyone who shouldn't have
access anymore.